requirements.txt
FogBugz
Cython
fastText
pyyaml